pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "MAP"
version = "0.1.0"
description = "Mini Amusement Parks"
authors = [
    { name = "Stephane Aroca-Ouellette", email = "stephane@skyfall.ai" }
]
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "matplotlib",
    "seaborn",
    "pygame",
    "pyperclip",
    "python-socketio",
    "line-profiler",
    "llm_model_cost",
    "gymnasium",
    "pyyaml",
    "pydantic",
    "requests",
    "websocket-client"
]

[project.optional-dependencies]

[tool.setuptools]
packages = ["map_py", "map_py.gui", "map_py.observations_and_actions", "map_py.tests"]

